Generate new internal component files
