+++ waspc/data/Generator/templates/react-app/prettier.config.js
@@ -1,0 +1,7 @@
+// Used for internal Wasp development only, not copied to generated app.
+module.exports = {
+  trailingComma: 'es5',
+  tabWidth: 2,
+  semi: false,
+  singleQuote: true,
+}
