--- waspc/data/Generator/templates/react-app/src/auth/forms/core/UserPassForm.tsx
@@ -1,5 +1,0 @@
-import { styled } from '../../../stitches.config'
-
-export const UserPassForm = styled('form', {
-    marginTop: '1.5rem'
-})
