+++ waspc/data/Generator/templates/react-app/src/auth/forms/internal/social/SocialIcons.tsx
@@ -8,14 +8,17 @@ const defaultStyles = css({
 export const Google = () => (
   <svg
     className={defaultStyles()}
-    aria-hidden="true" 
-    fill="currentColor" 
-    viewBox="0 0 24 24">
+    aria-hidden="true"
+    fill="currentColor"
+    viewBox="0 0 24 24"
+  >
     <g id="brand" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-        <g id="google" fill="#000000" fillRule="nonzero">
-            <path d="M11.99,13.9 L11.99,10.18 L21.35,10.18 C21.49,10.81 21.6,11.4 21.6,12.23 C21.6,17.94 17.77,22 12,22 C6.48,22 2,17.52 2,12 C2,6.48 6.48,2 12,2 C14.7,2 16.96,2.99 18.69,4.61 L15.85,7.37 C15.13,6.69 13.87,5.89 12,5.89 C8.69,5.89 5.99,8.64 5.99,12.01 C5.99,15.38 8.69,18.13 12,18.13 C15.83,18.13 17.24,15.48 17.5,13.91 L11.99,13.91 L11.99,13.9 Z" id="Shape">
-            </path>
-        </g>
+      <g id="google" fill="#000000" fillRule="nonzero">
+        <path
+          d="M11.99,13.9 L11.99,10.18 L21.35,10.18 C21.49,10.81 21.6,11.4 21.6,12.23 C21.6,17.94 17.77,22 12,22 C6.48,22 2,17.52 2,12 C2,6.48 6.48,2 12,2 C14.7,2 16.96,2.99 18.69,4.61 L15.85,7.37 C15.13,6.69 13.87,5.89 12,5.89 C8.69,5.89 5.99,8.64 5.99,12.01 C5.99,15.38 8.69,18.13 12,18.13 C15.83,18.13 17.24,15.48 17.5,13.91 L11.99,13.91 L11.99,13.9 Z"
+          id="Shape"
+        ></path>
+      </g>
     </g>
   </svg>
 )
@@ -23,10 +26,12 @@ export const Google = () => (
 export const GitHub = () => (
   <svg
     className={defaultStyles()}
-    aria-hidden="true" 
+    aria-hidden="true"
     fill="currentColor"
-    viewBox="0 0 20 20">
-    <path fillRule="evenodd"
+    viewBox="0 0 20 20"
+  >
+    <path
+      fillRule="evenodd"
       d="M10 0C4.477 0 0 4.484 0 10.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0110 4.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.203 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.942.359.31.678.921.678 1.856 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0020 10.017C20 4.484 15.522 0 10 0z"
       clipRule="evenodd"
     />
