+++ waspc/src/Wasp/Generator/WebAppGenerator/AuthG.hs
@@ -4,19 +4,16 @@ module Wasp.Generator.WebAppGenerator.AuthG
 where
 
 import Data.Aeson (object, (.=))
-import Data.Aeson.Types (Pair)
-import StrongPath (File', Path', Rel', reldir, relfile, (</>))
+import StrongPath (relfile)
 import Wasp.AppSpec (AppSpec)
 import qualified Wasp.AppSpec as AS
 import qualified Wasp.AppSpec.App as AS.App
 import qualified Wasp.AppSpec.App.Auth as AS.Auth
 import Wasp.AppSpec.Valid (getApp)
-import Wasp.Generator.AuthProviders (gitHubAuthProvider, googleAuthProvider)
-import qualified Wasp.Generator.AuthProviders.OAuth as OAuth
 import Wasp.Generator.Common (makeJsArrayFromHaskellList)
 import Wasp.Generator.FileDraft (FileDraft)
 import Wasp.Generator.Monad (Generator)
-import Wasp.Generator.WebAppGenerator.Auth.Common (getOnAuthSucceededRedirectToOrDefault)
+import Wasp.Generator.WebAppGenerator.Auth.AuthFormsG (genAuthForms)
 import Wasp.Generator.WebAppGenerator.Auth.EmailAuthG (genEmailAuth)
 import Wasp.Generator.WebAppGenerator.Auth.LocalAuthG (genLocalAuth)
 import Wasp.Generator.WebAppGenerator.Auth.OAuthAuthG (genOAuthAuth)
@@ -59,51 +56,3 @@ genUseAuth auth = return $ C.mkTmplFdWithData [relfile|src/auth/useAuth.ts|] tmp
   where
     tmplData = object ["entitiesGetMeDependsOn" .= makeJsArrayFromHaskellList [userEntityName]]
     userEntityName = AS.refName $ AS.Auth.userEntity auth
-
-genAuthForms :: AS.Auth.Auth -> Generator [FileDraft]
-genAuthForms auth =
-  sequence
-    [ genAuthForm auth,
-      copyTmplFile [relfile|auth/forms/Login.tsx|],
-      copyTmplFile [relfile|auth/forms/Signup.tsx|],
-      copyTmplFile [relfile|auth/forms/ResetPassword.tsx|],
-      copyTmplFile [relfile|auth/forms/ForgotPassword.tsx|],
-      copyTmplFile [relfile|auth/forms/VerifyEmail.tsx|],
-      copyTmplFile [relfile|auth/forms/types.ts|],
-      copyTmplFile [relfile|stitches.config.js|],
-      copyTmplFile [relfile|auth/forms/SocialIcons.tsx|],
-      copyTmplFile [relfile|auth/forms/SocialButton.tsx|]
-    ]
-  where
-    copyTmplFile = return . C.mkSrcTmplFd
-
-genAuthForm :: AS.Auth.Auth -> Generator FileDraft
-genAuthForm auth =
-  compileTmplToSamePath
-    [relfile|auth/forms/Auth.tsx|]
-    [ "onAuthSucceededRedirectTo" .= getOnAuthSucceededRedirectToOrDefault auth,
-      "areBothSocialAndPasswordBasedAuthEnabled" .= areBothSocialAndPasswordBasedAuthEnabled,
-      "isAnyPasswordBasedAuthEnabled" .= isAnyPasswordBasedAuthEnabled,
-      "isExternalAuthEnabled" .= AS.Auth.isExternalAuthEnabled auth,
-      -- Google
-      "isGoogleAuthEnabled" .= AS.Auth.isGoogleAuthEnabled auth,
-      "googleSignInPath" .= OAuth.serverLoginUrl googleAuthProvider,
-      -- GitHub
-      "isGitHubAuthEnabled" .= AS.Auth.isGitHubAuthEnabled auth,
-      "gitHubSignInPath" .= OAuth.serverLoginUrl gitHubAuthProvider,
-      -- Username and password
-      "isUsernameAndPasswordAuthEnabled" .= AS.Auth.isUsernameAndPasswordAuthEnabled auth,
-      -- Email
-      "isEmailAuthEnabled" .= AS.Auth.isEmailAuthEnabled auth,
-      "isEmailVerificationRequired" .= AS.Auth.isEmailVerificationRequired auth
-    ]
-  where
-    areBothSocialAndPasswordBasedAuthEnabled = AS.Auth.isExternalAuthEnabled auth && isAnyPasswordBasedAuthEnabled
-    isAnyPasswordBasedAuthEnabled = AS.Auth.isUsernameAndPasswordAuthEnabled auth || AS.Auth.isEmailAuthEnabled auth
-
-compileTmplToSamePath :: Path' Rel' File' -> [Pair] -> Generator FileDraft
-compileTmplToSamePath tmplFileInTmplSrcDir keyValuePairs =
-  return $
-    C.mkTmplFdWithData
-      (asTmplFile $ [reldir|src|] </> tmplFileInTmplSrcDir)
-      (object keyValuePairs)
